move tlds config outside of the test folder
